added experiment 1a
detect soma vs dendrite spike threshold as a function of distance from
the soma and frequency of synaptic activity.  Threshold is measured as
number of synapses required to generate a spike.

updated param.choose_seg_rand and param.set_weights:
the replace argument for choose_seg_rand now allows multiple selections
of the same segment.  If a segment is selected multiple times, the
weight in that segment is multiplied by the number repeats